refactor: 회원가입 페이지 컴포넌트 구조 변경
