+++ app/components/signup/index.ts
@@ -1,0 +1,2 @@
+export { default as SignupProfileForm } from './profile-form';
+export { default as SignupTermsForm } from './terms-form';
