+++ app/signup/profile/page.tsx
@@ -1,15 +1,17 @@
-'use client';
-
-import { SignupGowunTitle, SignupProfileForm } from '@/app/components';
+import { SignupProfileForm } from '@/app/components';
+import { gowunBatang } from '@/app/components/ui/fonts';
 
 function SignupProfilePage() {
   return (
-    <div className="bg-background h-full pb-13pxr flex flex-col justify-between gap-40pxr">
-      <SignupGowunTitle className="pt-47pxr">
+    <div className="h-full pb-13pxr flex flex-col justify-between gap-40pxr">
+      <p
+        className={`text-center ${gowunBatang.className} text-22pxr leading-[35px] tracking-[0.22px] pt-47pxr`}
+      >
         프로필 이름만 입력하면
         <br />
         가입이 완료돼요
-      </SignupGowunTitle>
+      </p>
+
       <SignupProfileForm />
     </div>
   );
